Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -8,7 +8,7 @@ int main()
 
     scanf("%d",&radius);
 
-    area = pi * (radius * radius);
+    area = pi * ((double)radius * radius);
 
     printf("Area: %.2lf",area);
